Created authentication methods for a user
Created the GeneratedSaltedHash method which takes in a plain text
password and a byte[] salt, hashes the salt and password together and
returns the byte[] of the hash.

Created CompareByteArrays which will be used to authenticate a users.
Takes in the hashed, salted password of a user and compares it with the
hashed, salted plain text password entered into the login form.

Added the CreateSalt method which takes in a size for the salt and
returns a Cryptographically random byte array.