Tweaked BroRequest code and partialview
Refined some of the code in the BroRequest region in ProfileController,
and edited some of the code so that it pulls the userid from the session
and gets the user from the dbcontext, rather than trying to pull the
user directly from the session.

Also, added accept and dismiss actionlinks to the brorequest
partialview.