Fixed Deleting/Creating Circles and removing users from circles
There is now a My Circles link in the header. The link takes you to a
page where you can view your circles and the users inside them. You
currently can create and delete circles, except for the original MyBros
circle, which is currently a static circle. Links for moving users
between circles and renaming circles are not yet in place.